requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
attrs==21.2.0
execnet==1.9.0
importlib-metadata==4.8.1
pytest-xdist==2.4.0
requests==2.22.0
selenium==3.14.1
toml==0.10.2
typing-extensions==3.10.0.2
webdriver-manager==3.3.0
zipp==3.6.0
Appium-Python-Client==1.1.0
pymysql==1.0.2
SQLAlchemy==1.4.35
cryptography==36.0.1
Faker==13.3.4